Update Python Database.py
This is the database code info for the databases. The "read in file" section can be optimized by creating a function or list to read in the files and then call the "Dataframe Creator" to create all the dataframes. However, it works so I'll work on optimizing the code later. CJH