CRQT-58: Added Histograms to SimulationResults.vue.
The first histogram is for the probability of DERIVED nodes, and the second is for the probability of ALL nodes. Currently, they use the same data, so one of them is just a placeholder.

Note, the histograms DO NOT support any number of bins other than 10. My next step is to fix that.